Obtain total scores (and more) of samba schools by year and/or school

get_parades() returns a tibble of total scores of samba schools. It can be filtered by year and/or school

Usage

get_parades(years, schools = "all")

Arguments

years

numeric: which years of parade to show. almost all years since 1932 are available.

schools

character: which samba schools to show. Case insensitive.

Value

A tibble. Its columns are school (samba school name), year

(parade year), school_total (samba school total score), school_rank

(samba school parade rank), school_remarks (remarks on the samba school parade), parade_location (samba school parade location), parade_date

(samba school parade date), parade_order (samba school parade order), parade_manager (parade manager), and theme (samba school parade theme). Each line corresponds to the total scores and parade info of a school in a given year.

About this data

This tibble has some confusing entries. One of the reasons: the competition changed a lot in the course of decades as perceived by the fluctuating number of schools, years where multiple competitions were held, and competitions with ties in the rank (when there was no criteria to break them).

Another reason are plain errors. Many entries have missing data. Some are highlighted as "Enredo não registrado" or "Resultado do desfile não registrado". Total scores may diverge from the final rank because of assessment errors.

These issues are rare or have low severity, meaning that the data is reliable. Although, one should pay attention to the parades remarks in the dataset itself and those from get_remarks() in order to get a better grasp of the happenings of a given year.

Examples

# Prints a table that shows the total scores and parade info of Portela and
# Estácio de Sá on the parades of 1968 and 1970
get_parades(years = c(1968, 1970), schools = c("portela", "Estácio de Sá"))
#> # A tibble: 4 × 10
#>   school      year schoo…¹ schoo…² schoo…³ parad…⁴ parad…⁵ parad…⁶ parad…⁷ theme
#>   <chr>      <int>   <dbl>   <dbl> <chr>   <chr>   <chr>     <dbl> <chr>   <chr>
#> 1 Portela     1968     103       4 Grupo 1 Candel… 25/02/…       5 AESEG … O tr…
#> 2 Estácio d…  1968      79       7 Grupo 1 Candel… 25/02/…       2 AESEG … Uma …
#> 3 Portela     1970      88       1 Grupo 1 Candel… 08/02/…      10 AESEG … Lend…
#> 4 Estácio d…  1970      72       7 Grupo 1 Candel… 08/02/…       5 AESEG … Terr…
#> # … with abbreviated variable names ¹​school_total, ²​school_rank,
#> #   ³​school_remarks, ⁴​parade_location, ⁵​parade_date, ⁶​parade_order,
#> #   ⁷​parade_manager