Obtain scores of samba schools by year, school and/or criterion

get_scores() returns a tibble of scores of samba schools. It can be filtered by year, school and/or criterion

Usage

get_scores(years, schools = "all", criterions = "all")

Arguments

years

numeric: which years of parade to show. almost all years since 1968 are available.

schools

character: which samba schools to show. Case insensitive.

criterions

character: which evaluation criteria to show. Case insensitive.

Value

A tibble. Its columns are school (samba school name), year

(parade year), score (samba school score), criteria (evaluation criteria), judge_name (judge name), and judge_number (judge id). Each line corresponds to a score given by a judge to a samba school on a criteria in a given year.

About this data

The competition changed a lot through the years, thus the list of schools and criteria, as well as the range of scores varies between years. The remarks from get_remarks() are a good source for better understanding the dynamics of the competition.

Examples

# Prints a table that shows the scores of Portela and
# Estácio de Sá on the parades of 1968 and 1970 on all criterions
get_scores(years = c(1968, 1970), schools = c("portela", "Estácio de Sá"))
#> # A tibble: 52 × 6
#>    school         year score criteria                 judge_name         judge…¹
#>    <chr>         <dbl> <dbl> <chr>                    <chr>              <glue> 
#>  1 Estácio de Sá  1968     6 ALEGORIAS E ADEREÇOS     Napoleão Muniz Fr… judge1 
#>  2 Portela        1968    10 ALEGORIAS E ADEREÇOS     Napoleão Muniz Fr… judge1 
#>  3 Estácio de Sá  1968     7 BATERIA                  João de Barros  B… judge1 
#>  4 Portela        1968     9 BATERIA                  João de Barros  B… judge1 
#>  5 Estácio de Sá  1968     7 COMISSÃO DE FRENTE       Danúbio Menezes G… judge1 
#>  6 Portela        1968     9 COMISSÃO DE FRENTE       Danúbio Menezes G… judge1 
#>  7 Estácio de Sá  1968     4 CONJUNTO                 Ítalo de Oliveira  judge1 
#>  8 Portela        1968     4 CONJUNTO                 Ítalo de Oliveira  judge1 
#>  9 Estácio de Sá  1968     4 DESFILE NÃO INTERROMPIDO Maurício Shermann  judge1 
#> 10 Estácio de Sá  1968     2 DESFILE NÃO INTERROMPIDO Sandra Dicken      judge2 
#> # … with 42 more rows, and abbreviated variable name ¹​judge_number